pyproject.toml
[project]
name = "airflow-dag-data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = []

[dependency-groups]
dev = ["apache-airflow==2.9.3", "pytest>=9.0.2", "ruff>=0.14.10"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
